Consolidate FailingFieldPlugin to a single class (#119302)
